Import modules directly instead of through barrels
`import { x } from "./resources"` made the importer depend on all 24 resource
modules, which is where most of garbo's import cycles came from. Naming the
defining module instead drops Rollup's circular dependency warnings from 61 to
29, and madge's from 127 to 41.

That left familiar/, outfit/, target/, resources/ and tasks/index.ts with no
consumers at all, so they're deleted.

Also moves logMessage/printLog into src/log.ts, so garboValue.ts — which sits
at the bottom of the stack — no longer imports lib.ts just for a print buffer.

The 29 warnings that remain are genuine mutual recursion in the valuation
model: turn estimates depend on fight counts and outfit choice, which depend
back on turn estimates. Breaking those needs a design decision rather than an
import rewrite.
