+++ packages/garbo/src/dailiespost.ts
@@ -1,4 +1,5 @@
-import { PostFreeFightQuest, runSafeGarboQuests } from "./tasks";
+import { runSafeGarboQuests } from "./tasks/engine";
+import { PostFreeFightQuest } from "./tasks/postFreeFight";
 
 export function postFreeFightDailySetup(): void {
   runSafeGarboQuests([PostFreeFightQuest]);
