+++ packages/garbo/src/diet.ts
@@ -96,12 +96,10 @@ import {
 import { acquire, priceCaps } from "./acquire";
 import { withVIPClan } from "./clan";
 import { globalOptions } from "./config";
-import {
-  beretEffectValue,
-  expectedGregs,
-  shouldAugustCast,
-  synthesize,
-} from "./resources";
+import { beretEffectValue } from "./resources/beret";
+import { expectedGregs } from "./resources/extrovermectin";
+import { shouldAugustCast } from "./resources/scepter";
+import { synthesize } from "./resources/synthesis";
 import {
   arrayEquals,
   HIGHLIGHT,
