--- packages/garbo/src/familiar/index.ts
@@ -1,11 +1,0 @@
-export {
-  canOpenRedPresent,
-  estimatedBarfExperience,
-  timeToMeatify,
-} from "./lib";
-export { meatFamiliar, setBestLeprechaunAsMeatFamiliar } from "./meatFamiliar";
-export { barfFamiliar } from "./barfFamiliar";
-export { freeFightFamiliar } from "./freeFightFamiliar";
-export { sandwormFamiliar } from "./sandwormFamiliar";
-export type { FamiliarMenuOptions } from "./freeFightFamiliar";
-export { getToyCupidBowFamiliars } from "./toyCupidBowFamiliar";
