+++ packages/garbo/src/lib.ts
@@ -354,18 +354,6 @@ export function kramcoGuaranteed(): boolean {
   );
 }
 
-const log: string[] = [];
-
-export function logMessage(message: string): void {
-  log.push(message);
-}
-
-export function printLog(color: string): void {
-  for (const message of log) {
-    print(message, color);
-  }
-}
-
 /**
  * Prints Garbo's help menu to the GCLI.
  */
