+++ packages/garbo/src/log.ts
@@ -1,0 +1,13 @@
+import { print } from "kolmafia";
+
+const log: string[] = [];
+
+export function logMessage(message: string): void {
+  log.push(message);
+}
+
+export function printLog(color: string): void {
+  for (const message of log) {
+    print(message, color);
+  }
+}
