--- packages/garbo/src/outfit/index.ts
@@ -1,12 +1,0 @@
-export { barfOutfit } from "./barf";
-export { meatTargetOutfit } from "./target";
-export { freeFightOutfit } from "./free";
-export type { FreeFightOutfitMenuOptions } from "./free";
-export {
-  familiarWaterBreathingEquipment,
-  waterBreathingEquipment,
-  useUPCs,
-  toSpec,
-  usingPurse,
-} from "./lib";
-export { bonusGear, magnifyingGlass } from "./dropsgear";
