+++ packages/garbo/src/post.ts
@@ -1,5 +1,5 @@
 import { safeRestore } from "./lib";
-import { runGarboQuests } from "./tasks";
+import { runGarboQuests } from "./tasks/engine";
 import { PostQuest } from "./tasks/post";
 
 export default function postCombatActions() {
