--- packages/garbo/src/resources/index.ts
@@ -1,24 +1,0 @@
-export * from "./scepter";
-export * from "./candyMap";
-export * from "./gingerbread";
-export * from "./worksheds";
-export * from "./extrovermectin";
-export * from "./realm";
-export * from "./synthesis";
-export * from "./latte";
-export * from "./aprilingband";
-export * from "./chestMimic";
-export * from "./mayam";
-export * from "./autumnaton";
-export * from "./stillsuit";
-export * from "./leprecondo";
-export * from "./yachtzee";
-export * from "./everfullDarts";
-export * from "./beret";
-export * from "./bloodCubicZirconia";
-export * from "./protonpack";
-export * from "./sealclub";
-export * from "./archaelogistSpade";
-export * from "./lucky";
-export * from "./doghouse";
-export * from "./clanVIP";
