+++ packages/garbo/src/tasks/buffExtension.ts
@@ -23,7 +23,8 @@ import {
   tryFindFreeRunOrBanish,
 } from "../lib";
 import { globalOptions } from "../config";
-import { freeFightOutfit, toSpec } from "../outfit";
+import { freeFightOutfit } from "../outfit/free";
+import { toSpec } from "../outfit/lib";
 import { Macro } from "../combat";
 import { GarboStrategy } from "../combatStrategy";
 import { effectExtenderValue } from "../potions";
