+++ packages/garbo/src/tasks/cockroach/prep.ts
@@ -30,11 +30,12 @@ import {
 import { acquire } from "../../acquire";
 import { Macro } from "../../combat";
 import { GarboStrategy } from "../../combatStrategy";
-import { freeFightFamiliar } from "../../familiar";
-import { freeFightOutfit, meatTargetOutfit } from "../../outfit";
+import { freeFightFamiliar } from "../../familiar/freeFightFamiliar";
+import { freeFightOutfit } from "../../outfit/free";
+import { meatTargetOutfit } from "../../outfit/target";
 import { GarboTask } from "../engine";
 import { bestCrewmate, dessertIslandWorthIt, outfitBonuses } from "./lib";
-import { doingGregFight } from "../../resources";
+import { doingGregFight } from "../../resources/extrovermectin";
 import { targetMeat, unignoreBeatenUp, userConfirmDialog } from "../../lib";
 import { globalOptions } from "../../config";
 import { DebuffPlanner } from "./debuffplanner";
