+++ packages/garbo/src/tasks/farm/dailyExtras.ts
@@ -1,4 +1,4 @@
-import { mayamCalendarSummon } from "../../resources";
+import { mayamCalendarSummon } from "../../resources/mayam";
 import { GarboTask } from "../engine";
 import { Quest } from "grimoire-kolmafia";
