+++ packages/garbo/src/tasks/freeGiantSandworm.ts
@@ -37,12 +37,12 @@ import { Macro } from "../combat";
 import { GarboStrategy } from "../combatStrategy";
 import { globalOptions } from "../config";
 import { garboValue } from "../garboValue";
-import { freeFightOutfit } from "../outfit";
+import { freeFightOutfit } from "../outfit/free";
 import { GarboTask } from "./engine";
 import { GarboFreeFightTask } from "./freeFight";
-import { sandwormFamiliar } from "../familiar";
+import { sandwormFamiliar } from "../familiar/sandwormFamiliar";
 import { sober } from "../lib";
-import { safeSweatBulletCasts } from "../resources";
+import { safeSweatBulletCasts } from "../resources/bloodCubicZirconia";
 import { acquire } from "../acquire";
 
 function sandwormSpec(spec: OutfitSpec = {}): OutfitSpec {
