--- packages/garbo/src/tasks/index.ts
@@ -1,15 +1,0 @@
-export { AscendingQuest } from "./ascending";
-export { CockroachSetup } from "./cockroach";
-export { DailyQuest } from "./daily";
-export { DailyFamiliarsQuest } from "./dailyFamiliars";
-export { DailyItemsQuest } from "./dailyItems";
-export { FarmQuests } from "./farm";
-export { FinishUpQuest } from "./finishUp";
-export { FreeFightQuest } from "./freeFight";
-export { FreeMimicEggDonationQuest } from "./freeEggDonation";
-export { PostFreeFightQuest } from "./postFreeFight";
-export { SetupTargetCopyQuest } from "./target";
-export { DailySeaQuest } from "./dailySea";
-export { PostQuest } from "./post";
-export { runGarboQuests, runSafeGarboQuests } from "./engine";
-export { EmbezzlerFightsQuest } from "./embezzler";
