+++ packages/garbo/src/tasks/post/index.ts
@@ -66,13 +66,10 @@ import { garboAverageValue } from "../../garboValue";
 import workshedTasks from "./worksheds";
 import { GarboPostTask } from "./lib";
 import { GarboTask } from "../engine";
-import {
-  autumnAtonManager,
-  hotTubAvailable,
-  lavaDogsAccessible,
-  lavaDogsComplete,
-  leprecondoTask,
-} from "../../resources";
+import { autumnAtonManager } from "../../resources/autumnaton";
+import { hotTubAvailable } from "../../resources/clanVIP";
+import { lavaDogsAccessible, lavaDogsComplete } from "../../resources/doghouse";
+import { leprecondoTask } from "../../resources/leprecondo";
 import { FarmingStrategy } from "../../farmingStrategy";
 
 const STUFF_TO_CLOSET = $items`bowling ball, funky junk key`;
