+++ packages/garbo/src/turns.ts
@@ -10,7 +10,7 @@ import { $familiar, $item, clamp, get, have } from "libram";
 import { globalOptions } from "./config";
 // Dumb circular import stuff
 import { usingThumbRing } from "./outfit/dropsgearAccessories";
-import { copyTargetCount } from "./target";
+import { copyTargetCount } from "./target/fights";
 import {
   digitizedMonstersRemainingForTurns,
   ESTIMATED_OVERDRUNK_TURNS,
